pyproject.toml
[tool.poetry]
name = "czech-tutor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pytelegrambotapi = "^4.18.1"
openai = "^1.30.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

